requirements.txt
Flask==2.3.2
PyYAML==6.0.1
pydantic==1.10.11
